requirements.txt
Django==2.2.9
django-import-export==2.0
django-storages==1.8
psycopg2-binary==2.8.4
whitenoise==5.0.1
Pillow==7.0.0
